# cmd/web/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CreateCatalogDefinitionRequest:
    properties:
      fields:
        items:
          $ref: '#/definitions/handlers.DefinitionFieldRequest'
        type: array
      pagination:
        $ref: '#/definitions/handlers.DefinitionPaginationRequest'
      productNavigation:
        $ref: '#/definitions/handlers.DefinitionNavigationRequest'
    type: object
  handlers.CreateProductDefinitionRequest:
    properties:
      fields:
        items:
          $ref: '#/definitions/handlers.DefinitionFieldRequest'
        type: array
    type: object
  handlers.CreateProductRequest:
    properties:
      name:
        type: string
      websites:
        items:
          type: string
        type: array
    type: object
  handlers.CreateProductResponse:
    properties:
      id:
        type: string
    type: object
  handlers.CreateScraperRequest:
    properties:
      cron:
        type: string
      urls:
        items:
          type: string
        type: array
    type: object
  handlers.CreateWebsiteRequest:
    properties:
      host:
        type: string
      name:
        type: string
    type: object
  handlers.CreateWebsiteResponse:
    properties:
      id:
        type: string
    type: object
  handlers.DefinitionFieldRequest:
    properties:
      displayName:
        type: string
      identifier:
        type: string
      required:
        type: boolean
      selector:
        type: string
    type: object
  handlers.DefinitionNavigationRequest:
    properties:
      fieldIdentifier:
        type: string
      navigate:
        type: boolean
    type: object
  handlers.DefinitionPaginationRequest:
    properties:
      maxPage:
        type: integer
      pageNumberParamName:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server for iTracker API.
  title: iTracker API
  version: "1.0"
paths:
  /products:
    post:
      consumes:
      - application/json
      operationId: create-product
      parameters:
      - description: CreateProductRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateProductResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new product
      tags:
      - Products
  /websites:
    post:
      consumes:
      - application/json
      operationId: create-website
      parameters:
      - description: CreateWebsiteRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateWebsiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateWebsiteResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new website
      tags:
      - Websites
  /websites/{id}/definitions/catalog:
    post:
      consumes:
      - application/json
      operationId: create-website-catalog-definition
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: CreateCatalogDefinitionRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCatalogDefinitionRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new catalog scraper definition for website
      tags:
      - Websites
  /websites/{id}/definitions/product:
    post:
      consumes:
      - application/json
      operationId: create-website-product-definition
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: CreateProductDefinitionRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateProductDefinitionRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new product scraper definition for website
      tags:
      - Websites
  /websites/{id}/scrapers/catalog:
    post:
      consumes:
      - application/json
      operationId: create-website-scraper
      parameters:
      - description: Website ID
        in: path
        name: id
        required: true
        type: string
      - description: CreateScraperRequest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateScraperRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "422":
          description: Unprocessable Entity
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new scraper for website
      tags:
      - Websites
swagger: "2.0"